Re-organizes apps into two separate entities.
