+++ src/main/java/CobSpecApp/CobSpecRoutes.java
@@ -1,6 +1,6 @@
 package CobSpecApp;
 
-import HTTPServer.*;
+import HTTPServer.Setup;
 
 import java.util.HashMap;
 import java.util.Map;
