+++ src/main/java/CobSpecApp/Handler.java
@@ -1,11 +1,10 @@
-package HTTPServer;
+package CobSpecApp;
+
+import HTTPServer.Response;
 
 import java.io.IOException;
 import java.util.Map;
 
-/**
- * Created by jphoenix on 8/23/16.
- */
 public interface Handler {
     Response handle(Map request) throws IOException;
 }
