+++ src/main/java/HTTPServer/HTTPService.java
@@ -1,5 +1,6 @@
 package HTTPServer;
 
+import CobSpecApp.Handler;
 import CobSpecApp.CobSpecRoutes;
 
 import java.io.*;
