+++ src/test/java/CobSpecApp/CobSpecRoutesTest.java
@@ -1,5 +1,5 @@
-import CobSpecApp.CobSpecRoutes;
-import HTTPServer.Handler;
+package CobSpecApp;
+
 import HTTPServer.Response;
 import HTTPServer.Setup;
 import junit.framework.TestCase;
