+++ src/test/java/CobSpecApp/OptionsHandlerTest.java
@@ -1,4 +1,5 @@
-import HTTPServer.OptionsHandler;
+package CobSpecApp;
+
 import HTTPServer.Response;
 import HTTPServer.Setup;
 import junit.framework.TestCase;
