+++ src/test/java/CobSpecApp/PostHandlerTest.java
@@ -1,4 +1,5 @@
-import HTTPServer.PostHandler;
+package CobSpecApp;
+
 import HTTPServer.Response;
 import HTTPServer.Setup;
 import junit.framework.TestCase;
