+++ src/test/java/Server/MockServerSocket.java
@@ -1,3 +1,5 @@
+package Server;
+
 import HTTPServer.ServerConnectible;
 import HTTPServer.Connectible;
