+++ src/test/java/Server/MockSocket.java
@@ -1,3 +1,5 @@
+package Server;
+
 import HTTPServer.Connectible;
 
 import java.io.*;
