+++ src/test/java/Server/ResponseTest.java
@@ -1,3 +1,5 @@
+package Server;
+
 import HTTPServer.Response;
 import junit.framework.TestCase;
